modification de la page recrutement

+++ src/app/main/body/pages/poste/poste.component.ts
@@ -1,8 +1,10 @@
-import { Component, OnInit } from '@angular/core';
+import { Component, computed, OnInit } from '@angular/core';
 import { Poste } from '../../../../api/models/Poste';
 import { PosteService } from '../../../../api/services/poste.service';
 import { FormBuilder, FormGroup } from '@angular/forms';
-import { MessageService } from 'primeng/api';
+import { ConfirmationService, MessageService } from 'primeng/api';
+import { PageEvent } from '../../../../api/models/pageEvent';
+import { SearchService } from '../../../../api/services/search.service';
 
 @Component({
   selector: 'app-poste',
@@ -13,18 +15,29 @@ export class PosteComponent implements OnInit {
   poste: Poste = {};
   posteForm: FormGroup;
   postes: Poste[] = [];
+  paginate: Poste[] = [];
+  first: number = 0;
+  rows: number = 10;
+  filteredData: any;
+
+  onPageChange(event: PageEvent) {
+    this.first = event.first ?? 0;
+    this.rows = event.rows ?? 10;
+    this.paginate = this.postes.slice(this.first, this.first + this.rows);
+  }
 
   constructor(
     private service: PosteService,
+    private searchService: SearchService,
     private formBuilder: FormBuilder,
-    private messageService: MessageService
+    private messageService: MessageService,
+    private confirmationService: ConfirmationService
   ) {
     this.posteForm = this.formBuilder.group({
       libelle: [''],
       niveauEtude: [''],
       description: [''],
       niveauDeSalaire: [0],
-      recruteur: [''],
       remarque: [''],
     });
   }
@@ -35,20 +48,23 @@ export class PosteComponent implements OnInit {
 
   addPoste() {
     this.service.createPoste(this.posteForm?.value).subscribe({
-      next: () => {
+      next: (data: Poste) => {
         this.messageService.add({
           severity: 'success',
           summary: 'Succès',
-          detail: 'Poste ajouté avec succès',
+          detail: `Poste ${data.libelle?.bold} ajouté avec succès`,
         });
         this.liste();
         this.posteForm?.reset();
       },
       error: (err) => {
+        // Extraction du message d'erreur personnalisé du backend
+        const errorMessage =
+          err.error?.error || "Erreur lors de l'ajout du poste";
         this.messageService.add({
           severity: 'error',
           summary: 'Erreur',
-          detail: "Erreur lors de l'ajout du poste",
+          detail: errorMessage,
         });
       },
     });
@@ -58,6 +74,18 @@ export class PosteComponent implements OnInit {
     this.service.getAllPostes().subscribe({
       next: (data: Poste[]) => {
         this.postes = data;
+        this.filteredData = computed(() => {
+          const query = this.searchService.searchQuery().toLowerCase();
+          const filtered = this.postes.filter((poste) =>
+            poste.libelle?.toLowerCase().includes(query)
+          );
+          const start = this.first;
+          const end = start + this.rows;
+
+          // Applique la pagination sur les résultats filtrés
+          return filtered.slice(start, end);
+        });
+        this.paginate = this.postes.slice(0, this.rows);
       },
       error: (err) => {
         this.messageService.add({
@@ -70,12 +98,36 @@ export class PosteComponent implements OnInit {
   }
 
   deletePoste(id: number) {
-    this.service.deletePoste(id).subscribe({
-      next: () => {
-        this.liste();
+    this.confirmationService.confirm({
+      header: 'Êtes-Vous sûr ?',
+      message: 'Veuillez confirmer la suppression',
+      accept: () => {
+        this.service.deletePoste(id).subscribe({
+          next: () => {
+            this.messageService.add({
+              severity: 'success',
+              summary: 'Succès',
+              detail: 'Poste supprimé avec succès',
+            });
+            this.liste();
+          },
+          error: (err) => {
+            const messageError =
+              err.error?.error || 'Oups! Un problème est survenu';
+            this.messageService.add({
+              severity: 'error',
+              summary: 'Erreur',
+              detail: messageError,
+            });
+          },
+        });
       },
-      error: (err) => {
-        console.log(err);
+      reject: () => {
+        this.messageService.add({
+          severity: 'warn',
+          summary: 'Alert',
+          detail: 'Vous avez annuler la suppression',
+        });
       },
     });
   }
@@ -86,30 +138,28 @@ export class PosteComponent implements OnInit {
       niveauEtude: poste.niveauEtude,
       description: poste.description,
       niveauDeSalaire: poste.niveauDeSalaire,
-      recruteur: poste.recruteur,
       remarque: poste.remarque,
     });
   }
 
   update() {
     if (this.poste.id)
-    this.service.updatePoste(this.poste.id, this.posteForm?.value).subscribe({
-      next: () => {
-        this.messageService.add({
-          severity: 'success',
-          summary: 'Succès',
-          detail: 'Poste modifié avec succès',
-        });
-        this.liste();
-        this.posteForm?.reset();
-      },
-      error: (err) => {
-        this.messageService.add({
-          severity: 'error',
-          summary: 'Erreur',
-          detail: "Erreur lors de la modification du poste",
-        });
-      },
-    });
+      this.service.updatePoste(this.poste.id, this.posteForm?.value).subscribe({
+        next: (data: Poste) => {
+          this.messageService.add({
+            severity: 'success',
+            summary: 'Succès',
+            detail: `Poste ${data.libelle?.bold} modifié avec succès`,
+          });
+          this.liste();
+        },
+        error: (err) => {
+          this.messageService.add({
+            severity: 'error',
+            summary: 'Erreur',
+            detail: 'Erreur lors de la modification du poste',
+          });
+        },
+      });
   }
 }
